+++ CHANGES.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+- [REMOVED] Removed oneway methods from ReplicatorFactory. Users should 
+  use the method described in the [replication documentation](https://github.com/cloudant/sync-android/blob/master/doc/replication.md) 
+  instead
+
 # 0.11.0 (2015-04-22)
 
 - [FIX] Using MongoDB query as the "gold" standard, query support for